# backend/modules/barberbooking/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API สำหรับโมดูล barberbooking",
        "title": "BarberBooking API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1/barberbooking",
    "paths": {
...
        "/tenants/:tenant_id/workinghour/branches/:branch_id": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "ดึงรายการ WorkingHour ทั้งหมดสำหรับสาขาที่ระบุ",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "WorkingHour"
                ],
                "summary": "ดึงเวลาเปิด-ปิดของสาขา",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "รหัสสาขา",
                        "name": "branch_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "คืนค่า status, message และ array ของ WorkingHour",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid branch ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "No working hours found for this branch",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to fetch working hours",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "รับรายการ WorkingHourInput หลายรายการเพื่อนำมาอัปเดตเวลาเปิด-ปิดของสาขาที่ระบุ (วันในสัปดาห์, เปิด/ปิด)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "WorkingHour"
                ],
                "summary": "อัปเดตเวลาเปิด-ปิดของสาขา (หลายวัน)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "รหัสสาขา",
                        "name": "branch_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Array ของ WorkingHourInput — ตัวอย่าง: [{\\",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/barberBookingDto.WorkingHourInput"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "คืนค่า status และข้อความยืนยันการอัปเดต",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid branch ID, ไม่มีข้อมูลใน body หรือ JSON ไม่ถูกต้อง",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Permission denied",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to update working hours",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "สร้าง WorkingHour ใหม่สำหรับสาขาที่ระบุ กำหนดวันในสัปดาห์ (0=อาทิตย์,...,6=เสาร์) และเวลาเปิด-ปิด",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "WorkingHour"
                ],
                "summary": "สร้างวันทำการใหม่ (เฉพาะ 1 วัน)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "รหัสสาขา",
                        "name": "branch_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Payload สำหรับสร้างวันทำการ (weekday, start_time, end_time)",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/barberBookingDto.WorkingHourInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "คืนค่า status และข้อความยืนยันการสร้าง",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid branch ID, weekday ไม่ถูกต้อง หรือ JSON ไม่ถูกต้อง",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Permission denied",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to create working hour",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "barberBookingDto.WorkingHourInput": {
            "type": "object",
            "properties": {
                "end_time": {
                    "description": "เวลาปิด",
                    "type": "string"
                },
                "start_time": {
                    "description": "เวลาเปิด",
                    "type": "string"
                },
                "weekday": {
                    "description": "0–6",
                    "type": "integer"
                }
            }
        },
...
    }
}